Stricter enforcement of enum options
