+++ src/cmd/vector_store/teardown.ts
@@ -5,10 +5,11 @@ import { getEnv, getEnvOrThrow } from '../../config';
 import { teardown as teardownChroma } from '../../vector_stores/chroma';
 import { teardown as teardownPinecone } from '../../vector_stores/pinecone';
 import { teardown as teardownPg } from '../../vector_stores/pgvector';
+import { SUPPORTED_VECTOR_STORES, type SupportedVectorStores } from '../../vector_stores';
 
 const argv = yargs(hideBin(process.argv))
   .option('store', {
-    type: 'string',
+    choices: SUPPORTED_VECTOR_STORES,
     description: 'The vector store',
     demandOption: true,
   })
@@ -16,7 +17,7 @@ const argv = yargs(hideBin(process.argv))
 
 teardown(argv.store);
 
-async function teardown(store: string) {
+async function teardown(store: SupportedVectorStores) {
   switch (store) {
     case 'chroma':
       return await teardownChroma({
@@ -29,7 +30,7 @@ async function teardown(store: string) {
         environment: getEnvOrThrow('PINECONE_ENVIRONMENT'),
         index: getEnvOrThrow('PINECONE_INDEX'),
       });
-    case 'pg':
+    case 'pgvector':
       return await teardownPg({
         tableName: getEnvOrThrow('PG_TABLE_NAME'),
         dsn: getEnvOrThrow('PG_DSN'),
