+++ src/cmd/vector_store/upload.ts
@@ -1,12 +1,13 @@
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 import { index } from '../../indexing';
-import { getVectorStore } from '../utils';
-import { getReader, SUPPORTED_READERS } from '../../readers';
+import { getReader, getVectorStore } from '../utils';
+import { SUPPORTED_READERS } from '../../readers';
+import { SUPPORTED_VECTOR_STORES } from '../../vector_stores';
 
 const argv = yargs(hideBin(process.argv))
   .option('store', {
-    type: 'string',
+    choices: SUPPORTED_VECTOR_STORES,
     description: 'The vector store',
     demandOption: true,
   })
