+++ src/vector_stores/index.ts
@@ -1,0 +1,6 @@
+import { NAME as chroma } from './chroma';
+import { NAME as pgvector } from './pgvector';
+import { NAME as pinecone } from './pinecone';
+
+export const SUPPORTED_VECTOR_STORES = [chroma, pgvector, pinecone] as const;
+export type SupportedVectorStores = (typeof SUPPORTED_VECTOR_STORES)[number];
